v0.31 Refactorizacion Permisos Localizacion
