+++ app/src/main/res/layout/style_user_show_info.xml
@@ -50,7 +50,6 @@
         android:backgroundTint="@android:color/transparent"
         android:contentDescription="Cerrar sesión"
         android:padding="16dp"
-        app:icon="@drawable/ic_baseline_logout_24"
         app:iconTint="@color/black"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
